scripts/requirements.txt
click
paho-mqtt
pycapnp
can
cantools
